requirements.txt
flask-restful==0.3.7
flask-mongoengine==0.9.5
gunicorn==19.9.0
passlib==1.7.1
pyjwt==1.7.1
flask-mail==0.9.1